+++ Plugins/Template/Server/DlgSettingsTemplateServer.h
@@ -1,0 +1,33 @@
+// Author: Kang Lin <kl222@126.com>
+
+#pragma once
+
+#include <QDialog>
+#include "ParameterServerUI.h"
+#include "ParameterFilterUI.h"
+
+namespace Ui {
+class CDlgSettingsTemplateServer;
+}
+
+class CParameterTemplateServer;
+class CDlgSettingsTemplateServer : public QDialog
+{
+    Q_OBJECT
+
+public:
+    explicit CDlgSettingsTemplateServer(CParameterTemplateServer* pPara,
+                                      QWidget *parent = nullptr);
+    virtual ~CDlgSettingsTemplateServer();
+
+    // QDialog interface
+public slots:
+    virtual void accept() override;
+
+private:
+    Ui::CDlgSettingsTemplateServer *ui;
+    CParameterTemplateServer* m_pPara;
+    CParameterServerUI* m_pServerUI;
+    CParameterFilterUI* m_pWhitelist;
+    CParameterFilterUI* m_pBlacklist;
+};